local/ollama_api/requirements.txt
fastapi==0.117.1
uvicorn[standard]==0.35.0
httpx==0.28.1
python-dotenv==1.1.1
